# openapi.yaml
openapi: 3.0.3
info:
  title: Gestion de Turnos
  description: 'App que permite gestionar turnos para evento unicos y recurrentes para diferentes organizaciones'
  contact:
    email: criscuelieduardo@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /evento:
    post:
      tags:
        - evento
      summary: Agrega un evento
      operationId: addEvent
      parameters:
        - name: cuit
          in: header
          description: cuit de la organizacion
          required: true
          schema:
            type: string
        - name: clave
          in: header
          description: clave de la organizacion
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '405':
          description: Entrada invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '412':
          description: Error de identificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
    put:
      tags:
        - evento
      summary: Modifica un evento
      operationId: updateEvent
      parameters:
        - name: nombreEvento
          in: header
          description: nombre del evento
          required: true
          schema:
            type: string
        - name: cuit
          in: header
          description: cuit de la organizacion
          required: true
          schema:
            type: string
        - name: clave
          in: header
          description: clave de la organizacion
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '404':
          description: Evento inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '405':
          description: Entrada invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '412':
          description: Error de identificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
    delete:
      tags:
        - evento
      summary: Elimina un evento
      operationId: deleteEvent
      parameters:
        - name: nombreEvento
          in: header
          description: nombre del evento
          required: true
          schema:
            type: string
        - name: cuit
          in: header
          description: cuit de la organizacion
          required: true
          schema:
            type: string
        - name: clave
          in: header
          description: clave de la organizacion
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '404':
          description: Evento inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '412':
          description: Error de identificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
  /evento/{cuit}/turnos:
    get:
      tags:
        - evento
      summary: Muestra todos los turnos de los eventos de la organizacion
      operationId: getTurns
      parameters:
        - name: cuit
          in: path
          description: cuit de la organizacion
          required: true
          schema:
            type: string
        - name: activo
          in: query
          description: estado del turno
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTurnos'          
        '404':
          description: No hay turnos registrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
...
components:
  schemas:
...
    Evento:
      type: object
      properties:
        nombre:
          type: string
        ubicacion:
          type: string
        fecha:
          type: string
        tipo:
          type: string
          enum:
            - UNICO
            - RECURRENTE
    Persona:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: number
...
    Turno:
      type: object
      properties:
        fecha:
          type: string
        hora:
          type: string
        persona:
          $ref: '#/components/schemas/Persona'
    Respuesta:
      type: object
      properties:
        status:
          type: number
        descripcion:
          type: string
...
    RespuestaTurnos:
      type: object
      allOf:
        - $ref: '#/components/schemas/Respuesta'
      properties:
        turnos:
          type: array
          items:  
            $ref: '#/components/schemas/Turno'
...